foreman: flatten CLI to a single `mngr foreman` command
Drop the `foreman` click group + `serve` subcommand in favor of one flat
`@click.command(name="foreman")` that serves directly, mirroring the
`mngr forward` plugin's structure. All flags (--host, --port, common opts)
are preserved; adds CommandHelpMetadata + pager help to match forward.

Collapses the `cli/` package (group.py, serve.py) into a single `cli.py`
module. Regenerates the CLI docs and updates the changelogs and README so
every `mngr foreman serve` reference becomes `mngr foreman`.
