+++ libs/mngr/changelog/minh-foreman-plugin.md
@@ -1,1 +1,1 @@
-- Registered the `mngr foreman` command in the generated CLI docs (`docs/commands/secondary/foreman.md`), documenting the foreman web remote-control plugin's `serve` subcommand.
+- Registered the `mngr foreman` command in the generated CLI docs (`docs/commands/secondary/foreman.md`), documenting the foreman web remote-control plugin — a single flat command that serves the web UI directly.
