+++ libs/mngr/docs/commands/secondary/foreman.md
@@ -2,20 +2,29 @@
 <!-- To modify, edit the command's help metadata and run: uv run python scripts/make_cli_docs.py -->
 
 # mngr foreman
-**Usage:**
+
+**Synopsis:**
 
 ```text
-mngr foreman [OPTIONS] COMMAND [ARGS]...
+mngr foreman [--host HOST] [--port PORT] [OPTIONS]
 ```
-**Options:**
 
+Always-on web remote control for your mngr agents [experimental].
 
-## mngr foreman serve
+Runs a single Flask server on this box, over every agent in
+mngr's view: a mobile-friendly chat UI for claude agents (live transcript with
+markdown, syntax highlighting, KaTeX, mermaid, inline images and file uploads;
+send messages; interrupt) and a web terminal (xterm.js over a pty bridge) for
+any agent type.
+
+No code is deployed to target boxes and there is no auth by design -- bind to a
+tailnet IP or firewall the port. Create agents with plain ``mngr create``;
+there is no foreman-specific create command or label filter.
 
 **Usage:**
 
 ```text
-mngr foreman serve [OPTIONS]
+mngr foreman [OPTIONS]
 ```
 **Options:**
 
@@ -33,6 +42,7 @@ mngr foreman serve [OPTIONS]
 | `--plugin`, `--enable-plugin` | text | Enable a plugin [repeatable] | None |
 | `--disable-plugin` | text | Disable a plugin [repeatable] | None |
 | `-S`, `--setting` | text | Override a config setting for this invocation (KEY=VALUE, dot-separated paths; append __extend to the leaf key to extend list/dict/set fields) [repeatable] | None |
+| `-h`, `--help` | boolean | Show this message and exit. | `False` |
 
 ## Other Options
 
@@ -40,3 +50,17 @@ mngr foreman serve [OPTIONS]
 | ---- | ---- | ----------- | ------- |
 | `--host` | text | Bind host (default from config, else 0.0.0.0). | None |
 | `--port` | integer | Bind port (default from config, else 8700). | None |
+
+## Examples
+
+**Serve on the default port**
+
+```bash
+$ mngr foreman --port 8700
+```
+
+**Bind to a specific tailnet IP**
+
+```bash
+$ mngr foreman --host 100.64.0.1
+```
