+++ libs/mngr_foreman/changelog/minh-foreman-plugin.md
@@ -1,1 +1,1 @@
-- Added `mngr_foreman`, an always-on web remote-control plugin for mngr agents. `mngr foreman serve` runs a single Flask server over every agent in mngr's view: a mobile-friendly chat UI for claude agents (live transcript with markdown, syntax highlighting, KaTeX, mermaid, inline images and file uploads; send messages; interrupt) and a web terminal (xterm.js over a pty bridge) for any agent type. No code is deployed to target boxes and there is no auth by design — bind to a private network. Includes a warm SSH connection pool that keeps resolution and connections hot for fast sends/reads.
+- Added `mngr_foreman`, an always-on web remote-control plugin for mngr agents. `mngr foreman` runs a single Flask server over every agent in mngr's view: a mobile-friendly chat UI for claude agents (live transcript with markdown, syntax highlighting, KaTeX, mermaid, inline images and file uploads; send messages; interrupt) and a web terminal (xterm.js over a pty bridge) for any agent type. No code is deployed to target boxes and there is no auth by design — bind to a private network. Includes a warm SSH connection pool that keeps resolution and connections hot for fast sends/reads.
